Extracted Layout from Movie Detail to external class.
